fix types

+++ src/main/java/dev/attara/stockify/application/service/impl/OrderServiceImpl.java
@@ -9,6 +9,7 @@ import dev.attara.stockify.domain.exception.NotAuthorizedException;
 import dev.attara.stockify.domain.exception.ProductNotFoundException;
 import dev.attara.stockify.domain.model.Order;
 import dev.attara.stockify.domain.model.Product;
+import dev.attara.stockify.domain.model.ProductLine;
 import dev.attara.stockify.domain.model.User;
 import dev.attara.stockify.domain.repository.OrderRepository;
 import dev.attara.stockify.domain.repository.ProductRepository;
@@ -20,7 +21,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -28,30 +28,25 @@ public class OrderServiceImpl implements OrderService {
 
     private final OrderRepository orderRepository;
 
-    private final ProductRepository productRepository;
-
     private final OrderMapper orderMapper;
+
+    private final ProductRepository productRepository;
 
     private final ProductLineMapper productLineMapper;
 
     @Override
     @Transactional
-    public OrderRecord createOrder(CreateOrderDTO createOrderDTO) {
-        List<ProductLineDTO> products = createOrderDTO.productLines();
+    public OrderRecord createOrder(CreateOrderDTO dto) {
         try {
-            // Process each product line
-            products.forEach(this::processProductLine);
-
-            // Create new order
-            Order order = Order.create(orderRepository.nextId(), createOrderDTO.user());
-            order.addProducts(products);
-
-            // Save order
+            List<ProductLine> productLines = dto.productLines()
+                    .stream()
+                    .map(productLineMapper::mapToDomain)
+                    .toList();
+            Order order = Order.create(orderRepository.nextId(), dto.user(),productLines);
+            productLines.forEach(this::processProductLine);
             orderRepository.save(order);
-
-            return orderMapper.toRecord(order);
+            return orderMapper.mapToRecord(order);
         } catch (ProductNotFoundException e) {
-            // Handle product not found error
             throw new RuntimeException("Failed to create order: Product not found", e);
         }
     }
@@ -59,28 +54,19 @@ public class OrderServiceImpl implements OrderService {
     @Override
     @Transactional
     public OrderRecord updateOrder(long orderId, List<ProductLineDTO> productLines, User user) throws NotAuthorizedException {
-        // Retrieve existing order
-        Order order = orderRepository.findById(orderId);
-
-        // Check if user is authorized to update the order
-        if (!order.getUser().equals(user)) {
-            throw new NotAuthorizedException();
-        }
-
         try {
-            // Process each product line
-            productLines.forEach(this::processProductLine);
-
-            // Update order with new product lines
-            order.addProducts(productLines);
-
-            // Save updated order
+            Order order = orderRepository.findById(orderId);
+            if (user.isNotAdmin() && !order.getUser().equals(user)) throw new NotAuthorizedException();
+            List<ProductLine> productLineList = productLines
+                    .stream()
+                    .map(productLineMapper::mapToDomain)
+                    .toList();
+            productLineList.forEach(this::processProductLine);
+            productLineList.forEach(order::addProductLine);
             orderRepository.save(order);
-
-            return orderMapper.toRecord(order);
+            return orderMapper.mapToRecord(order);
         } catch (ProductNotFoundException e) {
-            // Handle product not found error
-            throw new RuntimeException("Failed to update order: Product not found", e);
+            throw new RuntimeException("Failed to update order", e);
         }
     }
 
@@ -88,22 +74,29 @@ public class OrderServiceImpl implements OrderService {
     public List<OrderRecord> allOrders(User user) {
         List<Order> orders = orderRepository.findAll();
         if (user.isNotAdmin()) {
-            return orders.stream().filter(order -> order.getUser().equals(user)).map(orderMapper::toRecord).collect(Collectors.toList());
+            return orders.stream()
+                    .filter(order -> order.getUser().equals(user))
+                    .map(orderMapper::mapToRecord)
+                    .toList();
         }
-        return orders.stream().map(orderMapper::toRecord).collect(Collectors.toList());
+        return orders.stream()
+                .map(orderMapper::mapToRecord)
+                .toList();
     }
 
     @Override
     public List<OrderRecord> ordersByUserId(long userId) {
         List<Order> orders = orderRepository.findByUserId(userId);
-        return orders.stream().map(orderMapper::toRecord).collect(Collectors.toList());
+        return orders.stream()
+                .map(orderMapper::mapToRecord)
+                .toList();
     }
 
     @Override
     public OrderRecord getOrderById(long orderId, User user) {
         Order order = orderRepository.findById(orderId);
         if (user.isNotAdmin() && !order.getUser().equals(user)) return null;
-        return orderMapper.toRecord(order);
+        return orderMapper.mapToRecord(order);
     }
 
     @Override
@@ -112,23 +105,21 @@ public class OrderServiceImpl implements OrderService {
         if (user.isNotAdmin() && order.getUser().equals(user)) throw new NotAuthorizedException();
         return order.getProductLines()
                 .stream()
-                .map(productLineMapper::toRecord)
-                .collect(Collectors.toList());
+                .map(productLineMapper::mapToRecord)
+                .toList();
     }
 
     @Override
     public boolean deleteOrder(long orderId, User user) {
         Order order = orderRepository.findById(orderId);
+        if (user.isNotAdmin() && order.getUser().equals(user)) throw new NotAuthorizedException();
         orderRepository.delete(order);
         return true;
     }
 
-    private void processProductLine(@NonNull ProductLineDTO productLine) throws ProductNotFoundException {
-        Product product = productRepository.findById(productLine.productId());
-        if (product == null) {
-            throw new ProductNotFoundException(productLine.productId());
-        }
-        product.sell(productLine.quantity());
+    private void processProductLine(@NonNull ProductLine productLine) {
+        Product product = productLine.getProduct();
+        product.sell(productLine.getQuantity());
         productRepository.save(product);
     }
